There’s a web side of our site and an API version of our site, which we created in Episode 1.
The project is pretty simple, but awesome, you create programmers, give them a tag line, select from of our avatars and then battle a project. So we’ll start a battle here, fight against “InstaFaceTweet”... some magic happens, and boom! Our happy coder has defeated the InstaFaceTweet project.

So from an API perspective, you can see that we have a number of resources: we have programmers, we have projects and we also have battles. And they all relate to each other, which will be really important for episode 2.

What We’ve Already Accomplished

The API code all lives inside of the src/KnpU/CodeBattle/Controller/Api/ProgrammerController.php file. And in episode 1, we created endpoints for listing programmers, showing a single programmer, updating a programmer and deleting one:

// src/KnpU/CodeBattle/Controller/Api/ProgrammerController.php
// ...

class ProgrammerController extends BaseController {
  // ...

  public function newAction(Request {dollar_sign}request){
        // ...
  }

  // other methods for updating, deleting etc
}
We also used Behat to make some nice scenarios for this:

# features/api/programmer.feature
Feature: Programmer
  # ...

  Background:
    Given the user 'weaverryan' exists

  Scenario: Create a programmer
    Given I have the payload:
      '''
      {
        nickname: 'ObjectOrienter',
        'avatarNumber' : '2',
      }
      '''
    When I request 'POST /api/programmers'
    Then the response status code should be 201
    And the 'Location' header should be '/api/programmers/ObjectOrienter'

  # ... additional scenarios
This makes sure that our API doesn’t break. It also lets us design our API, before we think about implementing it.

Whenever I have an API, I like to have an object for each resource. Before we started in episode 1, I created a Programmer object. And this is actually what we’re allowing the API user to update, and we’re using this object to send data back to the user:

<?php

// src/KnpU/CodeBattle/Model/Programmer.php
// ...

class Programmer {
  public {dollar_sign}id;

  public {dollar_sign}nickname;

  public {dollar_sign}avatarNumber;

  public {dollar_sign}tagLine;

  // ... a few more properties
}

Serialization: Turning Objects into JSON or XML

So one of the key things we were doing was turning objects into JSON. For example, let’s look in showAction. We’re querying the database for a Programmer object, using a simple database-system I created in the background. Then ultimately, we pass that object into the serializeProgrammer function, which is a simple function we wrote inside this same class. It just takes a Programmer object and manually turns it into an array:

// src/KnpU/CodeBattle/Controller/Api/ProgrammerController.php
// ...

public function showAction({dollar_sign}nickname) {
  {dollar_sign}programmer = {dollar_sign}this->getProgrammerRepository()->findOneByNickname({dollar_sign}nickname);

    // ...
  {dollar_sign}data = {dollar_sign}this->serializeProgrammer({dollar_sign}programmer);
  {dollar_sign}response = new JsonResponse({dollar_sign}data, 200);

  return {dollar_sign}response;
}

// ...
private function serializeProgrammer(Programmer {dollar_sign}programmer) {
  return array(
    'nickname' => {dollar_sign}programmer->nickname,
    'avatarNumber' => {dollar_sign}programmer->avatarNumber,
    'powerLevel' => {dollar_sign}programmer->powerLevel,
    'tagLine' => {dollar_sign}programmer->tagLine,
  );
}
This transformation from an object into an array is really important because we’re going to be doing it for every resource across all of our endpoints.