Installing JMS Serializer The first thing we’re going to talk about is a library that makes this a lot easier and a lot more powerful, called a serializer. The one I like is called jms/serializer, so let’s Google for that. I’ll show you how this works as we start using it. But the first step to installing any library is to bring it in via Composer.
I’m opening a new tab, and going back into the root of my project, and then copying that command:

composer require jms/serializer

If you get a 'command not found' for composer, then you need to install it globally in your system(https://getcomposer.org/doc/00-intro.md#globally). Or, you can go and download it directly, and you’ll end up with a composer.phar file. In that case, you’ll run php composer.phar require instead.

Creating/Configuring the Serializer Object
While we’re waiting, let’s go back and look at the usage a little bit. What we’ll do is create an object called a “serializer”, and there’s this SerializerBuilder that helps us with this. Then we’ll pass it some data, which for us will be a Programmer object or a Battle object. And then it returns to you the actual JSON string. So it takes an object and turns it into a string:

// from the serialization documentation
{dollar_sign}serializer = JMS\Serializer\SerializerBuilder::create()->build();
{dollar_sign}jsonContent = {dollar_sign}serializer->serialize(, 'json');
echo {dollar_sign}jsonContent; // or return it in a Response