Using the Serializer Object
The important thing here is that we have a serializer object and we can access it from any of our controllers. Let’s open our ProgrammerController and rename serializeProgrammer to just serialize since it can serialize anything.

I’ve setup my application so that I can reference any of those global objects by saying ->container['serializer']. This will look different for you: the important point is that we need to access that object we just configured.

Now, just call serialize() on it, just like the documentation told us. I’ll put json as the second argument so we get JSON. The serializer can also give us another format, like XML.

Perfect! Now let’s look to see where we used the serializeProgrammer function before. That old function returned an array, but serialize now returns the JSON string. So now we can return a normal Response object and just pass the JSON string that we want. The one thing we’ll lose is the Content-Type header being set to application/json, but we’ll fix that in a second.

Let’s go and make similar changes to the rest of our code.

In fact, when we have the collection of Programmer objects, things get much easier. We can pass the entire array of objects and it’s smart enough to know how to serialize that. You can already start to see some of the benefits of using a serializer.

Compared with what we had before, not a lot should have changed, because the serializer should give us a JSON structure with all the properties in Programmer. That’s practically the same as we were doing before.

So let’s run our tests!

php vendor/bin/behat

We’ve totally changed how a Programmer gets turned into JSON, but almost every test passes already! We’ll debug that failure next.