Serializer Configuration (SerializationContext) Time to tackle that test failure! You can see that there is one strange problem. The 'GET one programmer' scenario says it’s expecting a tagLine property, but you can see that the JSON response has a lot of other fields, but not tagLine. I know what the issue is, so we’ll fix it in a second.
But first, one thing the test didn’t show yet, is that we’re losing the Content-Type of application/json. We need to centralize the logic that creates our Response as much as possible so that all of our responses are very very consistent.

For example, right now - we’re creating the Response in every controller, and so every controller method is responsible for remembering to set the Content-Type header. That’s pretty error-prone, which means I am guaranteed to mess it up.