Central Method for Creating all API Responses
Instead, let’s go to BaseController, the parent class of ProgrammerController. Let’s create a couple of new functions. First, back in ProgrammerController, cut the serialize function, move it into the BaseController and change it to be protected.
Now, when we have multiple controllers in the future, we can just re-use this method to serialize other resources, like Battles.

Second, create a new function called createApiResponse with 2 arguments: the data and the status code. The data could be a Programmer a Battle or anything else. Then we’ll let it call the serialize function. And finally, create the Response and make sure the Content-Type header is set perfectly