Back in ProgrammerController, we can simplify a lot of things. Let’s search for new Response, because we can replace these. In newAction we can say  = ->createApiResponse() and pass it the  object and the 201 status code. And we can still add any other headers we need.
I’ll copy this code and change the other spots in this controller. The deleteAction returns a 204, which is the blank response. So there’s no need to use this fancy serialization stuff here.

Now let’s try the tests again, so we can make sure we see just that same one failure:

php vendor/bin/behat
And we do!