To Serialize or Not Serialize Null Values?
This failure is caused by something specific to the serializer. In this test, the programmer doesn’t actually have a tagLine - we could see this if we looked in the database:

# features/api/programmer.feature
# ...

Scenario: GET one programmer
  Given the following programmers exist:
    | nickname   | avatarNumber |
    | UnitTester | 3            |
  When I request 'GET /api/programmers/UnitTester'
  Then the response status code should be 200
  And the following properties should exist:
    '''
    nickname
    avatarNumber
    powerLevel
    tagLine
    '''
  And the 'nickname' property should equal 'UnitTester'
When the serializer sees null values, it has 2 options: return the property with a null value, or omit the property entirely. By default, the serializer actually omits null properties. We’ll fix this, because I like always having the same fields on a response, even if some of the data is empty.

Go back to the serialize function. To configure the serializer, create a SerializationContext object. Next, call setSerializeNull and pass true. Now, pass this context as the 3rd argument to serialize.

There’s not a lot you can customize in this way, but serializeNull happens to be one of them.

Back to the tests!

php vendor/bin/behat
Boom, everything passes! We’ve changed to use the serializer, and it’s now taking care of all of the heavy-lifting for us. This will be really powerful as we serialize more and more objects.