Alright, now let me show you one other powerful thing about the serializer. The control you have over your objects! By default, it serializes all of your properties. But let’s say we don’t want the userId to be serialized. After all this isn’t really an important field for the API, it’s just the id of the user that created the programmer.
To start, open up the feature file and add a line to test that this property isn’t in the response.

We’re using a custom Behat context that I created for this project with a lot of these nice sentences. To see a list of all of them, you can run behat with the -dl option, which stands for 'definition list':

php vendor/bin/behat -dl
Try running the tests again. It should fail, and it does - saying that the userId property should not exist, but we can see that it in fact is there.