Configuring the Serializer with Annotations
As soon as you want to take control over what properties are returned, we’re going to use annotations. Let’s look at their documentation first and find the Annotation Reference section - this is by far the most interesting page. The first item on the list is what we need, but there’s a huge list of annotations that give you all sorts of control.

Remember, whenever you use an annotation, you need a use statement for it. PHPStorm is tagging in to help me auto-complete the ExclusionPolicy class. Then I’ll remove the last part and alias this to Serializer. This will allow us to use any of the JMS serializer annotations by starting with @Serializer.

For example, on top of the class, we can say @Serializer\ExclusionPolicy('all'). We’ve now told the serializer to not serialize any of the properties in this class, unless we tell it to specifically. Whereas before, it was serializing everything.

To actually include things, we whitelist them with the @Serializer\Expose annotation. I’ll copy this and use it on nickname, avatarNumber, tagLine and powerLevel fields.

This is just one of the customizations you can make with annotations.

Now let’s re-run the test:

php vendor/bin/behat
Success! This time the userId is not returned in our JSON.

If you want to know more, check out that annotation reference section. But we’re also going to do more in the next videos.