Authenticating a User via a Token
So let’s hook this up! Some of this is specific to Silex’s security system, but in case you’re using something else, we’ll stay high level enough to see what types of things you need to do in your system to make it happen. As always, if you have questions, just ask them in the comments!

Inside this Security/ directory here, I’ve already set up a bunch of things for an API token authentication system

1) ApiTokenListener: Gets the Token from the Request
The first thing we’re going to do is open this ApiTokenListener. I’ve written some fake code in here as you can see. The job of the listener is to look at the request object and get the token information off of it. And hey, since we’re sending the token on the Authorization header, we are going to look for it there. So let’s get rid of this hard coded text and instead go get that Authorization header. You can say {dollar_sign}request->headers->get('Authorization'). That’s going to get you the actual raw token ABCD123 type of thing.