The only other thing that could have gone wrong, was if the user  not logged in at all and they hit updateAction. Then, we would die inside this function. The problem was that {dollar_sign}this->getLoggedInUser would be null and we’ld call the id property on a null object.
So before we call this function, we need to make sure the user is at least logged in. If they aren’t, then they are definitely not the owner of this programmer.

So let’s create another function here called enforceUserSecurity. In this case, go back to ProgrammerController and grab the logic right in BaseController.

There we go. And from inside enforceProgrammerOwnershipSecurity we can just make sure that the user is actually logged in.

And in ProgrammerController, we can do the same thing and save ourselves a little bit of code.

Between these two new methods, we have a really easy way to go function by function inside of our controller to make sure that we’re enforcing the right type of security.

Because we’re sending our Authorization header in the background of our scenarios, we should be able to run our entire programmer.feature and see it pass.