Authentication Error Format It’s finally time to create a scenario to check and see what happens if we send an invalid token. So let’s do that right now.
In this case, I’m not going to add a user to the database with this token. I’m just going to send a token that doesn’t exist.

Then, we’ll just make any request that requires authentication. The response status code should be 401 and remember we’re always returning that API problem format that has a detail property on it. And here, we can say whatever we want. To be nice to the users, let’s set it to 'Invalid Credentials' so they know what went wrong. It’s not like they forgot the token, it just wasn’t valid.

Let’s try this out. Again we can run Behat on one particular scenario. This one starts on line 11:

php vendor/bin/behat features/api/authentication.feature:11
In fact you can see it almost passed, so out of the box things are working. We are denying access, sending a 401, and because of our security error handling in that ApiEntryPoint class, we’re sending a nice api problem format with the actual detail set to 'Invalid Credentials.' Like before, this message comes from deep inside Silex and is describing what’s going wrong.

We’ll see the difference this makes. We’re expecting 'Invalid Credentials!' and we are getting it with a period.