Returning application/problem+json for Security Errors
Next, we need all of our errors to always return that same API problem response format. And when we return this format, we should always send back its special Content-Type so let’s make sure it’s correct.

Ahh! It’s not coming back with that. We are getting an application/problem-like format, but without the right Content-Type header. It’s coming back as a simple application/json.