In our app, when an exception is thrown, there are 2 different places that take care of things. Most errors are handled in the Application class. We added this in episode 1. But security errors are handled in ApiEntryPoint, and it’s responsible for returning some helpful response:
// src/KnpU/CodeBattle/Security/Authentication/ApiEntryPoint.php
// ...

public function start(Request {dollar_sign}request, AuthenticationException {dollar_sign}authException = null) {
  {dollar_sign}message = {dollar_sign}this->getMessage({dollar_sign}authException);

  {dollar_sign}response = new JsonResponse(array('detail' => {dollar_sign}message), 401);

  return {dollar_sign}response;
}
So for example here, you can see why we get the detail and why we get the 401. If I change this to 403, this proves that this class is responsible for the error responses. Let’s add the application/problem+json Content-Type header