Using the ApiProblem Class For Security Errors
For consistency, one of the things we did in Episode 1 is actually create an ApiProblem class. The idea was whenever you had some sort of error response you needed to send back, you could create this ApiProblem object, which will help you structure things and avoid typos in any keys.

Right now inside of the ApiEntryPoint, we’re kind of creating the API problem structure by hand, which is something I don’t want to do. Let’s leverage our ApiProblem class instead.

So first, I’m closing a couple of these classes. Inside ApiProblem there is a type property. The spec document that describes this format says that we should have a type field and that it should be a unique string for each error in your application. Right now we have two: validation_error as one unique thing that can go wrong and invalid_body_format as another.

That’s if the client sends us json, but the json is malformed. Now we have a third type of error, which is when you send us bad credentials. So let’s add a new constant here called authentication_error. And I’m just making up this string, it’s not terribly important. And then down here is a map from those types to a human readable text that will live on the title key.

The purpose of this is that when we create a new ApiProblem, we are forced to pass in a type and then that has a nice little map to the title. So given a certain type, you always get this nice same identical human readable explanation for it. You don’t have to duplicate the titles all around your codebase.