Back in ApiEntryPoint, instead of this stuff, you can create a new ApiProblem object. Add our use statement for that. The status code we know is 401 and the type is going to be our new authentication_error type.
So it’s a nice way to make sure we don’t just invent new types all over the place.

And then, we set the detail. The detail is going to be the message that comes from Silex whenever something goes wrong related to security. Based on what went wrong, we will get a different message here and we can use the translator to control it.

Then down here for the response, we can say just new JsonResponse. For the content, we can say {dollar_sign}problem->toArray(). This is a function we used earlier: it just takes all those properties and turns them into an array. Now we’ll use {dollar_sign}problem->getStatusCode(). And we’ll keep the response headers already set.

So this is a small improvement. I’m more consistent in my code, so my API will be more consistent too. If I need to create an api problem response, I won’t do it by hand. The ApiProblem class does some special things for us, attaching the title and making sure we have a few defined types. If we try this, we should get the same result as before and we do. Perfect.