Let’s go further! Even the creation of the response is error prone. So right now, in both the Application class where we have our error handler and inside of this ApiEntryPoint, we create the JsonResponse and we set the Content-Type header to application/problem+json by hand. I don’t want to have a lot of these laying around: I want to go through one central spot.
The fix for this has nothing to do with Silex or API’s: we’re just going to do a bit of refactoring and repeat ourselves a little bit less.

Lets create a new PHP class called APIProblemResponseFactory and its job will be to create API problem responses.
So we’ll create a single function called createResponse and it will take in an ApiProblem object and create the response for that. And most of this we can just copy from our error handler code.
I’ll make sure that I add a couple of use statements here. Perfect, it takes in the ApiProblem, transforms that into json, and makes sure that the Content-Type header is set. So if we can use this instead of repeating that logic elsewhere, it’s going to save us some trouble.