Creating an api.response_factory Service
Like we saw before, inside of Silex there is a way to create global objects called services. We did this for the serializer, which let us use it in multiple places. So I’m going to do the same thing with the api.response_factory. And we’ll just return new ApiProblemResponseFactory. Of course, like anything else don’t forget to add the use statement for that.

Yes, this class is getting a little crazy. And that’s it!

Down inside this class we’ll use that key to access the object and make use of it. I have that same {dollar_sign}app variable, so I can get rid of all this stuff here. Pass the ApiProblem object to createResponse and there we go!