Injecting ApiProblemResponseFactory into ApiEntryPoint
We can do the same thing inside the ApiEntryPoint. I need to practice a little bit of dependency injection, and if this is a new idea to you or going over your head, we have a free tutorial about dependency injection. I highly recommend you check it out, it’s going to change the way you code.

So in Application, I’m going to find the entry point and I’m actually going to go past that new factory object right to it as the second argument to the __construct function of our ApiEntryPoint.
This means here I will now have a second argument. Don’t forget the use statement for that and we’ll just set that on a new property.

So now, when this object is created we’re going to have access to this ApiProblemResponseFactory. Down below, we can just use it.
So we still create the ApiProblem object, but I don’t want to do any of this other stuff. And that’s it! We just reduced duplication, let’s try our tests:

php vendor/bin/behat features/api/authentication.feature
Those pass too! Let’s try all of our tests for the programmer.

php vendor/bin/behat features/api/programmer.feature
Sahweet! They’re passing too! So there’s no chance of duplication because everything is going through that same class.