Activating Silex’s HTTP Basic Authentication In our controller we need to check to see if the password is correct for the user. But hey, let’s not do that, Silex can help us with some of this straightforward logic. In my main Application class, where I configure my security, I’ve already setup things to allow http basic to happen. By adding this little key here, when the http basic username and password come into the request, the Silex security system will automatically look up the user object and deny access if they have the wrong password.
It’s kind of like our API token system: but instead of sending a token it’s going to be reading it off of the HTTP Basic username and password headers. That’s pretty handy.

That means that in the controller, if we need the actual user object we don’t need to query for it - the security system already did that for us. We can just say {dollar_sign}this->getLoggedInUser().

We don’t really know if the user logged in via HTTP basic or passed a token, and frankly we don’t care. And since we need our user to be logged in for this endpoint, we can use our nice {dollar_sign}this->enforceUserSecurity() function.

Perfect, let’s try that out. And it passes with almost no effort!